Add Files page to dashboard
